requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
datasets>=2.19.0
jiwer>=3.0.0
librosa>=0.10.0
torchaudio>=2.0.0
soundfile>=0.12.0
fastapi>=0.110.0
uvicorn>=0.29.0
langdetect>=1.0.9
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
huggingface_hub>=0.22.0
accelerate>=0.29.0
evaluate>=0.4.0
wandb>=0.16.0
bitsandbytes>=0.43.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
